pyproject.toml
[project]
name = "adhd-study-guide-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.43.0",
    "boto3>=1.35.0",
    "chromadb>=0.6.8",
    "datasets>=4.8.4",
    "google-genai>=0.2.0",
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
    "textstat>=0.7.3",
]

[tool.ruff]
exclude = [".venv", "data/make_txt.py"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
